+++ api/v1/api-doc.js
@@ -239,10 +239,6 @@ const apiDoc = {
                 allowedTools: {
                     type: "string",
                 },
-                forbiddenTools: {
-                    type: "array",
-                    items: {},
-                },
                 blender: {
                     $ref: "#/definitions/Blender",
                 },