+++ api/v1/paths/admin/sources/{id}.js
@@ -1,11 +1,22 @@
-const { sourceModel } = require("../../../../model/sources");
-const userManager = require("../../../../bin/user.");
+const { sourceModel } = require("../../../../../model/sources");
+const userManager = require("../../../../../bin/user.");
 
 module.exports = function () {
     let operations = {
+        GET,
         DELETE,
         PUT,
     };
+
+    async function GET(req, res, _next) {
+        const userInfo = await userManager.getUser(req.user);
+        const source = await sourceModel.getOneUserSource(userInfo.user, req.params.id);
+        if (source) {
+            res.status(200).json(source);
+            return;
+        }
+        res.status(400).json({ message: `Source with id ${req.params.id} not found` });
+    }
 
     async function DELETE(req, res, next) {
         if (!req.params.id) {
@@ -19,8 +30,7 @@ module.exports = function () {
                 res.status(500).json({ message: `I couldn't delete resource ${sourceIdToDelete}. Maybe it has been deleted already?` });
                 return;
             }
-            const userInfo = await userManager.getUser(req.user);
-            const sources = await sourceModel.getOwnedSources(userInfo.user);
+            const sources = await sourceModel.getAllSources();
             res.status(200).json({ message: `${sourceIdToDelete} successfully deleted`, resources: sources });
         } catch (err) {
             next(err);
@@ -40,14 +50,27 @@ module.exports = function () {
                 res.status(400).json({ message: "Resource does not exist. If you want to create another resource, use POST instead." });
                 return;
             }
-            const userInfo = await userManager.getUser(req.user);
-            const sources = await sourceModel.getOwnedSources(userInfo.user);
+            const sources = await sourceModel.getAllSources();
             res.status(200).json({ message: `${sourceIdToUpdate} successfully updated`, resources: sources });
         } catch (err) {
             next(err);
         }
     }
 
+    GET.apiDoc = {
+        summary: "This resource returns profiles list or a profile if an id is provided",
+        operationId: "getProfiles",
+        security: [{ restrictLoggedUser: [] }],
+        parameters: [],
+        responses: {
+            200: {
+                description: "Profiles",
+                schema: {
+                    $ref: "#/definitions/Profile",
+                },
+            },
+        },
+    };
     DELETE.apiDoc = {
         summary: "Delete a specific source",
         security: [{ restrictAdmin: [] }],