+++ mainapp/src/Profile.ts
@@ -87,8 +87,7 @@ type ProfileJson = {
     name: string;
     allowedSourceSchemas: string[];
     sourcesAccessControl: Record<string, SourceAccessControl>;
-    allowedTools: string[] | string;
-    forbiddenTools: string[];
+    allowedTools: string[];
     theme?: string;
 };
 
@@ -100,7 +99,6 @@ const decodeProfile = (key: string, profile: ProfileJson): Profile => {
         allowedSourceSchemas: profile.allowedSourceSchemas,
         sourcesAccessControl: profile.sourcesAccessControl,
         allowedTools: profile.allowedTools,
-        forbiddenTools: profile.forbiddenTools,
         theme: profile.theme,
     };
 };
@@ -110,7 +108,7 @@ export type Profile = z.infer<typeof ProfileSchema>;
 const SourceAccessControlSchema = z.union([z.literal("forbidden"), z.literal("read"), z.literal("readwrite")]);
 
 const ProfileSchema = z.object({
-    name: z.string(),
+    name: z.string().default(""),
     _type: z.string().optional(),
     theme: z.string().optional(),
     allowedSourceSchemas: z
@@ -118,8 +116,7 @@ const ProfileSchema = z.object({
         .nullish()
         .transform((a) => (a ?? []).flatMap((item) => (item ? item : []))),
     sourcesAccessControl: z.record(SourceAccessControlSchema).default({}),
-    allowedTools: z.union([z.string(), z.array(z.string())]).default("ALL"),
-    forbiddenTools: z.array(z.string()).default([]),
+    allowedTools: z.array(z.string()).default([]),
     id: z.string().default(ulid()),
 });
 
@@ -141,8 +138,7 @@ export const defaultProfile = (uuid: string): Profile => {
         id: uuid,
         allowedSourceSchemas: [],
         sourcesAccessControl: {},
-        allowedTools: "ALL",
-        forbiddenTools: [],
+        allowedTools: [],
     };
 };
 export { getProfiles, deleteProfile, ProfileSchema };