+++ public/vocables/modules/shared/export.js
@@ -450,4 +450,3 @@ fixedColumns: true*/
 export default Export;
 
 window.Export = Export;
-