+++ public/vocables/modules/tools/UserManagement.js
@@ -1,17 +1,25 @@
-const UserManagement = (function () {
+const UserManagement = (() => {
     const self = {};
-    self.onSourceSelect = function () {};
+    self.createApp = null;
+    self.mountApp = null;
+    import("/assets/user_management.js");
 
-    self.onLoaded = function () {
+    self.onSourceSelect = () => {};
+
+    self.onLoaded = () => {
         $("#accordion").accordion("option", { active: 2 });
-        setTimeout(function () {
-            import("/assets/user_management.js");
-            $("#graphDiv").html("");
-            $("#graphDiv").html(`
-                    <div id="mount-user-management-here"></div>
-                `);
+
+        setTimeout(() => {
+            $("#mainDialogDiv").dialog("option", "title", "User Management");
+            $("#mainDialogDiv").on("dialogclose", (event, ui) => self.unload());
+            $("#mainDialogDiv").html(`<div style="width:90vw;height:80vh"><div id="mount-user-management-here"></div></div>`);
+            $("#mainDialogDiv").dialog("open");
+
+            self.mountApp = self.createApp();
         }, 200);
     };
+
+    self.unload = () => self.mountApp();
 
     return self;
 })();